refactor: moving rollback functions into rollback package.
