+++ configlib/structs.go
@@ -42,6 +42,7 @@ type PkgrConfig struct {
 	Packages       []string            `yaml:"Packages,omitempty"`
 	Suggests       bool                `yaml:"Suggests,omitempty"`
 	Repos          []map[string]string `yaml:"Repos,omitempty"`
+	Rollback	   bool				   `yaml:"Rollback,omitempty"`
 	Library        string              `yaml:"Library,omitempty"`
 	LibPaths       []string            `yaml:"LibPaths,omitempty"`
 	Customizations Customizations      `yaml:"Customizations,omitempty"`
