+++ rollback/operations_test.go
@@ -119,3 +119,54 @@ func (suite *OperationsTestSuite) TestRollbackPackageEnvironment_PackagesAreCase
 	suite.True(afero.DirExists(suite.FileSystem, filepath.Join(suite.FilePrefix, "test-library", "crayon")))
 }
 */
+
+// Have to use an OS filesystem for this because of this issue: https://github.com/spf13/afero/issues/141
+func (suite *OperationsTestSuite) TestRollbackUpdatePackages_RestoresWhenNoActiveInstallation() {
+
+	_ = suite.FileSystem.MkdirAll(filepath.Join(suite.FilePrefix, "test-library", "__OLD__CatsAndOranges"), 0755)
+	_, _ = suite.FileSystem.Create(filepath.Join(suite.FilePrefix, "test-library", "__OLD__CatsAndOranges", "DESCRIPTION"))
+
+	updateAttemptFixture := []UpdateAttempt{
+		UpdateAttempt{
+			Package:                "CatsAndOranges",
+			BackupPackageDirectory: filepath.Join(suite.FilePrefix, "test-library", "__OLD__CatsAndOranges"),
+			ActivePackageDirectory: filepath.Join(suite.FilePrefix, "test-library", "CatsAndOranges"),
+			NewVersion:             "2",
+			OldVersion:             "1",
+		},
+	}
+	RollbackUpdatePackages(suite.FileSystem, updateAttemptFixture)
+
+	suite.True(afero.DirExists(suite.FileSystem, filepath.Join(suite.FilePrefix, "test-library", "CatsAndOranges")))
+	suite.True(afero.Exists(suite.FileSystem, filepath.Join(suite.FilePrefix, "test-library", "CatsAndOranges", "DESCRIPTION")))
+	suite.False(afero.DirExists(suite.FileSystem, filepath.Join(suite.FilePrefix, "test-library", "__OLD__CatsAndOranges")))
+	suite.False(afero.Exists(suite.FileSystem, filepath.Join(suite.FilePrefix, "test-library" ,"__OLD__CatsAndOranges","DESCRIPTION")))
+
+}
+
+// Have to use an OS filesystem for this because of this issue: https://github.com/spf13/afero/issues/141
+func (suite *OperationsTestSuite) TestRollbackUpdatePackages_OverwritesFreshInstallation() {
+
+	_ = suite.FileSystem.MkdirAll(filepath.Join(suite.FilePrefix, "test-library", "CatsAndOranges"), 0755)
+	_, _ = suite.FileSystem.Create(filepath.Join(suite.FilePrefix, "test-library", "CatsAndOranges", "DESCRIPTION_New"))
+	_ = suite.FileSystem.MkdirAll(filepath.Join(suite.FilePrefix, "test-library", "__OLD__CatsAndOranges"), 0755)
+	_, _ = suite.FileSystem.Create(filepath.Join(suite.FilePrefix, "test-library", "__OLD__CatsAndOranges", "DESCRIPTION"))
+
+	updateAttemptFixture := []UpdateAttempt{
+		UpdateAttempt{
+			Package:                "CatsAndOranges",
+			BackupPackageDirectory: filepath.Join(suite.FilePrefix, "test-library", "__OLD__CatsAndOranges"),
+			ActivePackageDirectory: filepath.Join(suite.FilePrefix, "test-library", "CatsAndOranges"),
+			NewVersion:             "2",
+			OldVersion:             "1",
+		},
+	}
+	RollbackUpdatePackages(suite.FileSystem, updateAttemptFixture)
+
+	suite.True(afero.DirExists(suite.FileSystem, filepath.Join(suite.FilePrefix, "test-library", "CatsAndOranges")))
+	suite.True(afero.Exists(suite.FileSystem, filepath.Join(suite.FilePrefix, "test-library", "CatsAndOranges", "DESCRIPTION")))
+	suite.False(afero.Exists(suite.FileSystem, filepath.Join(suite.FilePrefix, "test-library", "CatsAndOranges", "DESCRIPTION_New")))
+	suite.False(afero.DirExists(suite.FileSystem, filepath.Join(suite.FilePrefix, "test-library", "__OLD__CatsAndOranges")))
+	suite.False(afero.Exists(suite.FileSystem, filepath.Join(suite.FilePrefix, "test-library" ,"__OLD__CatsAndOranges","DESCRIPTION")))
+
+}
