+++ rollback/structs.go
@@ -1,0 +1,11 @@
+package rollback
+
+// UpdateAttempt ...
+type UpdateAttempt struct {
+	Package                string
+	ActivePackageDirectory string
+	BackupPackageDirectory string
+	OldVersion             string
+	NewVersion             string
+}
+
